bugfixes & optimisations
